[158] Remove redundant chessmove2 file
