+++ lambda/chessgame/src/compute.rs
@@ -5,7 +5,7 @@ use bytes::Bytes;
 use rusoto_core::Region;
 use rusoto_lambda::{InvocationRequest, Lambda, LambdaClient};
 
-use lambda_payloads::chessmove2::{ChooseMoveEvent, ChooseMoveOutput};
+use lambda_payloads::chessmove::{ChooseMoveEvent, ChooseMoveOutput};
 use myopic_brain::anyhow::{anyhow, Result};
 
 pub struct MoveLambdaClient {
