+++ lambda/payloads/src/lib.rs
@@ -1,4 +1,3 @@
 pub mod benchmark;
 pub mod chessgame;
 pub mod chessmove;
-pub mod chessmove2;
